Create initial draft version of Test view in Grading.
