+++ bundles/edu.kit.kastel.sdq.eclipse.common.api/src/edu/kit/kastel/sdq/eclipse/grading/api/artemis/mapping/ResultsDTO.java
@@ -2,12 +2,16 @@
 package edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping;
 
 import java.io.Serializable;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
 import java.util.Date;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class ResultsDTO implements Serializable {
 	private static final long serialVersionUID = 6637703343535347213L;
+
+	private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 
 	@JsonProperty
 	public int id;
@@ -28,5 +32,10 @@ public class ResultsDTO implements Serializable {
 
 	public ResultsDTO() {
 		// NOP
+	}
+
+	public String toLocalDateTimeString() {
+		var time = this.completionDate == null ? null : this.completionDate.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime();
+		return time == null ? "" : time.format(formatter);
 	}
 }
