+++ bundles/edu.kit.kastel.sdq.eclipse.common.api/src/edu/kit/kastel/sdq/eclipse/grading/api/client/websocket/WebsocketCallback.java
@@ -8,29 +8,23 @@ public interface WebsocketCallback {
 
 	/**
 	 * Handles events send over result topic.
-	 * 
+	 *
 	 * @param send payload via topic
 	 */
 	void handleResult(Object payload);
 
 	/**
 	 * Handles events send over submission topic.
-	 * 
+	 *
 	 * @param send payload via topic
 	 */
 	void handleSubmission(Object payload);
 
 	/**
-	 * Handles general websocket errors.
-	 * 
+	 * Handles websocket errors.
+	 *
 	 * @param the thrown exception.
 	 */
 	void handleException(Throwable e);
 
-	/**
-	 * Handles transport error.
-	 * 
-	 * @param the thrown exception.
-	 */
-	void handleTransportError(Throwable e);
 }
