+++ bundles/edu.kit.kastel.sdq.eclipse.common.api/src/edu/kit/kastel/sdq/eclipse/grading/api/controller/IArtemisController.java
@@ -9,6 +9,8 @@ import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.ICourse;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.IExam;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.IExercise;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.ISubmission;
+import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.ResultsDTO;
+import edu.kit.kastel.sdq.eclipse.grading.api.util.Pair;
 
 /**
  * Works as an interface from backend to REST-clients
@@ -20,7 +22,7 @@ public interface IArtemisController extends IController {
 	 * @return all IFeedbacks that were gotten in the process of locking the given
 	 *         submission.
 	 */
-	List<Feedback> getAllFeedbacksGottenFromLocking(ISubmission submission);
+	Pair<ResultsDTO, List<Feedback>> getAllFeedbacksGottenFromLocking(ISubmission submission);
 
 	/**
 	 *
