+++ bundles/edu.kit.kastel.sdq.eclipse.common.api/src/edu/kit/kastel/sdq/eclipse/grading/api/controller/IStudentArtemisController.java
@@ -2,7 +2,6 @@
 package edu.kit.kastel.sdq.eclipse.grading.api.controller;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.Feedback;
@@ -12,6 +11,7 @@ import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.IExercise;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.IStudentExam;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.ParticipationDTO;
 import edu.kit.kastel.sdq.eclipse.grading.api.artemis.mapping.ResultsDTO;
+import edu.kit.kastel.sdq.eclipse.grading.api.util.Pair;
 
 /**
  * Works as an interface from backend to REST-clients. It handles specific tasks
@@ -21,7 +21,7 @@ public interface IStudentArtemisController extends IArtemisController {
 
 	/**
 	 * Fetches the exercises of the exam with the title examTitle.
-	 * 
+	 *
 	 * @param examTitle
 	 * @return StudentExam with exam and all its exercises.
 	 */
@@ -29,16 +29,16 @@ public interface IStudentArtemisController extends IArtemisController {
 
 	/**
 	 * Fetches the feedback of all results from the given exercise.
-	 * 
+	 *
 	 * @param course
 	 * @param excerise
 	 * @return Map from result to its feedbacks.
 	 */
-	Map<ResultsDTO, List<Feedback>> getFeedbackForExercise(ICourse course, IExercise excerise);
+	Pair<ResultsDTO, List<Feedback>> getFeedbackForExercise(ICourse course, IExercise excerise);
 
 	/**
 	 * Starts an exam.
-	 * 
+	 *
 	 * @param course
 	 * @param exam
 	 * @return the exam and its exercises.
@@ -47,7 +47,7 @@ public interface IStudentArtemisController extends IArtemisController {
 
 	/**
 	 * Returns participation if exists of given exercise and current user.
-	 * 
+	 *
 	 * @param course
 	 * @param exercise
 	 * @return
